Simplify XML schema by flattening properties
I noticed that some of the plants had a class element with properties
inside it, and others had top-level properties with dot-namespaced
names. This led me to realized that the class elements were just
another way of writing the "class" part of the property names. This
mean we can convert all of the subproperties to top-level properties
without changing their meaning!

I confirmed the flattened file generated by this schema/tool works
with Alpha 13.6. If it ceases to work in the future (unlikely, since
they have data in it both ways already), then I'll just have to add a
postprocessing step to the save function.

If we convert the properties before loading them into the DataSet,
this removes recursion from the schema, removes the class attribute
from properties, and means that the name and value attributes are now
required. It also means that every property has a block_id (no DBNull
values!).

Theoretically, the xsd.exe-generated DataSet code will now work, since
there are no recursive tables. That may come in the future. The method
of loading the schema from a resource works for now.

I also took the liberty of changing the data type for //block@id to
something more semantically correct. Also both parameters for //block
are required, and //block@id is a unique key. It is not the primary
key (block_id); that is autogenerated by the DataSet.