Build out the application as far as possible using a DataSet
* Finish changing over to to using the strongly typed DataSet. This
  includes using custom columns for the DataGridView
* Flattening the XML breaks repairing, so unflatten it (except for
  Map.Color, which must be flattened).
* Sort the properties and drops for easier manual editing.
* Simplify the UI for what is reasonable to implement (well, except for
  the actions group). Any read/write exposing of properties requires
  manual event handling, again defeating the purpose of the "database".

At this point, I believe I have reached the limits of the "pretend the
XML is a database" model. To do any significant processing of blocks,
they need to be linked to materials. It's not possible to load multiple
XML documents with schemas into a DataSet, so there's no way to have a
foreign key relationship between blocks and materials, so I'm back to
iteration and string comparison anyway. So I'm going to implement a
custom class or three and parse the XML myself.

While the application does not modify the data in the XML right now, it
does serve as a decent formatter for making the rather messy shipped
file more modification-friendly.

I let the designer do what it wanted with the generated code. It is
going to be rewritten soon anyway.